+++ LibraWallet/Vendors/LibraKit/LibraMultiHDWallet/LibraMultiHDWallet.swift
@@ -9,7 +9,7 @@ import Foundation
 import CryptoSwift
 struct LibraSeedAndDepth {
     var seed:[UInt8]
-    var depth: Int
+    var depth: UInt64
     var sequence: Int
 }
